Fixed a bug related to transcribing files multiple times. A subsequent transcription would trigger transcription of already transcribed files. Also refactored the App.js file to improve readability.
